Changed some stuff regarding parsing request parameters
We had the bugs in the parsing function that were hard to debug because of the unstandardized way we were parsing each parameter. I changed some things to try to make it so they were all handled the same way. Specifically, I wanted each function to take the index of the first character of the parameter and return with the index pointing to one past the delimeter (ie the start of the next parameter).

I moved some string parsing into another function to fit this interface. I did make some assumptions to write these functions so we will have to add in some careful input checking to make sure our server doesn't crash on badly formed inputs. I will say breaking the server was equally as likely with the old way of parsing the input, so no backwards progress was made.

author: Mark Koester